refactor part 3: move updateGoogleProjectIam from GoogleServicesDAO to setupGoogleProject in WorkspaceService
