+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/GoogleServicesDAO.scala
@@ -35,9 +35,7 @@ abstract class GoogleServicesDAO(groupsPrefix: String) extends ErrorReportable {
                      bucketName: String,
                      labels: Map[String, String],
                      parentSpan: Span = null,
-                     bucketLocation: Option[String],
-                     policyMap: Map[SamResourcePolicyName, WorkbenchEmail],
-                     billingProjectOwnerPolicyEmail: WorkbenchEmail): Future[GoogleWorkspaceInfo]
+                     bucketLocation: Option[String]): Future[GoogleWorkspaceInfo]
 
   def getGoogleProject(googleProject: GoogleProjectId): Future[Project]
