+++ core/src/test/scala/org/broadinstitute/dsde/rawls/webservice/ApiServiceSpec.scala
@@ -36,7 +36,7 @@ import org.broadinstitute.dsde.rawls.status.StatusService
 import org.broadinstitute.dsde.rawls.user.UserService
 import org.broadinstitute.dsde.rawls.util.MockitoTestUtils
 import org.broadinstitute.dsde.rawls.workspace.WorkspaceService
-import org.broadinstitute.dsde.workbench.google.mock.MockGoogleBigQueryDAO
+import org.broadinstitute.dsde.workbench.google.mock.{MockGoogleBigQueryDAO, MockGoogleIamDAO}
 import org.scalatest.concurrent.Eventually
 import spray.json._
 
@@ -222,7 +222,10 @@ trait ApiServiceSpec extends TestDriverComponentWithFlatSpecAndMatchers with Raw
       requesterPaysSetupService,
       entityManager,
       resourceBufferService,
-      resourceBufferSaEmail
+      resourceBufferSaEmail,
+      googleIamDao = new MockGoogleIamDAO,
+      googleProjectOwnerRole = "fakeGoogleProjectOwnerRole",
+      googleProjectViewerRole = "fakeGoogleProjectViewerRole"
     )_
 
     override val entityServiceConstructor = EntityService.constructor(
